working on Event.vue

+++ src/components/Event.vue
@@ -1,13 +1,16 @@
 <template>
   <div class="component">
     <button v-on:click="this.$router.push('/')" class="button"> X </button>
-    <div class="component">
+      <div>
     <h1>{{state.event.name}}</h1>
     <p>{{state.event.eventDate}}</p>
     <p>Anzahl Personen: {{state.event.participants.length}}</p>
       <br>
+        <div>
+          <button v-on:click="this.$router.replace('/person/new')" class="button" > Add someone + </button>
+        </div>
       <div>
-        <button v-on:click="this.deleteEvent" class="button" > Löschen </button>
+        <button v-on:click="this.deleteEvent" class="button" > Event furtputze </button>
       </div>
     </div>
   </div>
@@ -37,29 +40,33 @@ export default {
     await REST_interface.getCollection("events").then(resp=>{
       let eventDb = resp.filter(item => item._id === this.$route.params._id);
       let newDate = new Date(eventDb[0].event.eventDate);
-      eventDb[0].event.eventDate = "📆 " + new moment(newDate).format('L') + "\n";
+      eventDb[0].event.eventDate = "📆 " + new moment(newDate).format('LL') + "\n";
       this.state.event = eventDb[0].event;
     }).catch(err=>{
       console.log('Event load failed: ' + err);
-      this.$router.push('/');
+      this.$router.replace('/');
     });
     await REST_interface.getCollection("persons").then(resp=>{
       this.state.participants = resp.filter(person => person.event === this.state.event._id);
     }).catch(err=>{
       console.log('Event load failed: ' + err);
-      this.$router.push('/');
+      this.$router.replace('/');
     });
   },
   methods:{
     async deleteEvent() {
-      await REST_interface.deleteItemInCollection("events",this.$route.params._id).then(()=>{
-        this.$router.push('/');
-      }).catch(err=>{
-        alert('Event deleting failed: ' + err)
-        console.log('Event delete failed: ' + err);
-        this.$router.push('/');
-      });
-    }
+      if (confirm('Are you sure you want to delete this?')) {
+        await REST_interface.deleteItemInCollection("events",this.$route.params._id).then(()=>{
+          this.$router.replace('/');
+        }).catch(err=>{
+          alert('Event deleting failed: ' + err)
+          console.log('Event delete failed: ' + err);
+          this.$router.push('/');
+        });
+      } else {
+        alert("You're weak... weichbächer")
+      }
+    },
   }
 }
 </script>
@@ -67,9 +74,14 @@ export default {
 <style scoped>
 .component{
   color: white;
-  margin: 1em;
+  background-color: #2c3e50;
+  padding: 20px;
+  margin-left: 50px;
+  margin-right: 50px;
+  border-radius: 2em;
 }
 .button{
+  margin-top: 10px;
   background-color: transparent;
   padding: 10px 40px 10px 40px;
   border-radius: 2em;
